Clean remoteParticipants state in the hook
